execinfrapb: break dependency on parser
This commit breaks dependency of `sql/execinfrapb` package on
`sql/parser` package. It does the following mechanical changes:
- `colinfo.CompareDatums` is moved towards its single call site
- `init*FromAST` methods of window frame are moved to `sql`
- `MakeEvalContext` is moved (and renamed) towards its single call site
- `ExprHelper` and friends is moved into a new `execinfra/execexpr`
package (we didn't want to move it into `execinfra` to not introduce
a dependency of `colexecagg` on `execinfra`)
- `ExprFmtCtxBase` is moved towards its single call site.

All changes are no-ops (modulo some renaming and unexporting).

Release note: None
