+++ pkg/BUILD.bazel
@@ -476,6 +476,7 @@ ALL_TESTS = [
     "//pkg/sql/doctor:doctor_test",
     "//pkg/sql/enum:enum_test",
     "//pkg/sql/evalcatalog:evalcatalog_disallowed_imports_test",
+    "//pkg/sql/execinfra/execexpr:execexpr_test",
     "//pkg/sql/execinfra:execinfra_disallowed_imports_test",
     "//pkg/sql/execinfra:execinfra_test",
     "//pkg/sql/execinfrapb:execinfrapb_disallowed_imports_test",
@@ -1981,6 +1982,8 @@ GO_TARGETS = [
     "//pkg/sql/enum:enum_test",
     "//pkg/sql/evalcatalog:evalcatalog",
     "//pkg/sql/execinfra/execagg:execagg",
+    "//pkg/sql/execinfra/execexpr:execexpr",
+    "//pkg/sql/execinfra/execexpr:execexpr_test",
     "//pkg/sql/execinfra/execopnode:execopnode",
     "//pkg/sql/execinfra/execreleasable:execreleasable",
     "//pkg/sql/execinfra:execinfra",
