+++ pkg/sql/BUILD.bazel
@@ -427,6 +427,7 @@ go_library(
         "//pkg/sql/evalcatalog",
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfra/execagg",
+        "//pkg/sql/execinfra/execexpr",
         "//pkg/sql/execinfra/execopnode",
         "//pkg/sql/execinfra/execreleasable",
         "//pkg/sql/execinfrapb",
