+++ pkg/sql/catalog/colinfo/ordering.go
@@ -7,9 +7,7 @@ package colinfo
 
 import (
 	"bytes"
-	"context"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 )
@@ -63,27 +61,3 @@ func (ordering ColumnOrdering) String(columns ResultColumns) string {
 
 // NoOrdering is used to indicate an empty ColumnOrdering.
 var NoOrdering ColumnOrdering
-
-// CompareDatums compares two datum rows according to a column ordering. Returns:
-//   - 0 if lhs and rhs are equal on the ordering columns;
-//   - less than 0 if lhs comes first;
-//   - greater than 0 if rhs comes first.
-func CompareDatums(
-	ctx context.Context, ordering ColumnOrdering, evalCtx *eval.Context, lhs, rhs tree.Datums,
-) int {
-	for _, c := range ordering {
-		// TODO(pmattis): This is assuming that the datum types are compatible. I'm
-		// not sure this always holds as `CASE` expressions can return different
-		// types for a column for different rows. Investigate how other RDBMs
-		// handle this.
-		if cmp, err := lhs[c.ColIdx].Compare(ctx, evalCtx, rhs[c.ColIdx]); err != nil {
-			panic(err)
-		} else if cmp != 0 {
-			if c.Direction == encoding.Descending {
-				cmp = -cmp
-			}
-			return cmp
-		}
-	}
-	return 0
-}
