+++ pkg/sql/execinfra/processorsbase.go
@@ -10,6 +10,7 @@ import (
 	"math"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/execinfra/execexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
@@ -76,7 +77,7 @@ type DoesNotUseTxn interface {
 type ProcOutputHelper struct {
 	// eh only contains expressions if we have at least one rendering. It will
 	// not be used if outputCols is set.
-	eh execinfrapb.MultiExprHelper
+	eh execexpr.MultiHelper
 	// outputCols is non-nil if we have a projection. Only one of renderExprs and
 	// outputCols can be set. Note that 0-length projections are possible, in
 	// which case outputCols will be 0-length but non-nil.
@@ -165,7 +166,8 @@ func (h *ProcOutputHelper) Init(
 		if evalCtx == flowCtx.EvalCtx {
 			// We haven't created a copy of the eval context, and we have some
 			// renders, then we'll need to create a copy ourselves since we're
-			// going to use the ExprHelper which might mutate the eval context.
+			// going to use the execexpr.Helper which might mutate the eval
+			// context.
 			evalCtx = flowCtx.NewEvalCtx()
 		}
 		if cap(h.OutputTypes) >= nRenders {
