+++ pkg/sql/execinfrapb/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "api.go",
         "component_stats.go",
         "data.go",
-        "expr.go",
         "flow_diagram.go",
         "processors.go",
         "processors_changefeed.go",
@@ -24,27 +23,16 @@ go_library(
         "//pkg/roachpb",
         "//pkg/rpc/rpcbase",
         "//pkg/security/username",
-        "//pkg/sql/catalog",
         "//pkg/sql/catalog/catalogkeys",
         "//pkg/sql/catalog/catenumpb",
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/fetchpb",
-        "//pkg/sql/catalog/typedesc",
-        "//pkg/sql/parser",
-        "//pkg/sql/pgwire/pgcode",
-        "//pkg/sql/pgwire/pgerror",
-        "//pkg/sql/rowenc",
         "//pkg/sql/sem/catid",  # keep
-        "//pkg/sql/sem/eval",
-        "//pkg/sql/sem/normalize",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sem/tree/treewindow",
-        "//pkg/sql/sessiondata",
-        "//pkg/sql/types",
         "//pkg/util",
         "//pkg/util/buildutil",
-        "//pkg/util/duration",
         "//pkg/util/encoding",
         "//pkg/util/humanizeutil",
         "//pkg/util/optional",
@@ -65,7 +53,6 @@ go_test(
     size = "small",
     srcs = [
         "component_stats_test.go",
-        "expr_test.go",
         "flow_diagram_external_test.go",
         "flow_diagram_test.go",
     ],
@@ -75,15 +62,11 @@ go_test(
         "//pkg/keys",
         "//pkg/roachpb",
         "//pkg/security/username",
-        "//pkg/settings/cluster",
         "//pkg/sql",
         "//pkg/sql/catalog/catpb",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/fetchpb",
         "//pkg/sql/rowenc",
-        "//pkg/sql/sem/eval",
-        "//pkg/sql/sem/tree",
-        "//pkg/sql/types",
         "//pkg/util/leaktest",
         "//pkg/util/optional",
         "@com_github_stretchr_testify//require",
@@ -160,5 +143,8 @@ go_proto_library(
 
 disallowed_imports_test(
     "execinfrapb",
-    ["//pkg/sql/sem/builtins"],
+    [
+        "//pkg/sql/parser",
+        "//pkg/sql/sem/builtins",
+    ],
 )
