+++ pkg/sql/execinfrapb/api.go
@@ -9,8 +9,6 @@ import (
 	"strconv"
 
 	"github.com/cockroachdb/cockroach/pkg/security/username"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
-	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 )
 
@@ -40,19 +38,6 @@ func (f FlowID) Equal(other FlowID) bool {
 // IsUnset returns whether the FlowID is unset.
 func (f FlowID) IsUnset() bool {
 	return f.UUID.Equal(uuid.Nil)
-}
-
-// MakeEvalContext serializes some of the fields of a eval.Context into a
-// execinfrapb.EvalContext proto.
-func MakeEvalContext(evalCtx *eval.Context) EvalContext {
-	sessionDataProto := evalCtx.SessionData().SessionData
-	sessiondata.MarshalNonLocal(evalCtx.SessionData(), &sessionDataProto)
-	return EvalContext{
-		SessionData:                       sessionDataProto,
-		StmtTimestampNanos:                evalCtx.StmtTimestamp.UnixNano(),
-		TxnTimestampNanos:                 evalCtx.TxnTimestamp.UnixNano(),
-		TestingKnobsForceProductionValues: evalCtx.TestingKnobs.ForceProductionValues,
-	}
 }
 
 // User accesses the user field.
