+++ pkg/sql/execinfrapb/processors.go
@@ -6,18 +6,11 @@
 package execinfrapb
 
 import (
-	context "context"
 	"unicode/utf8"
 	"unsafe"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catenumpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
-	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree/treewindow"
-	"github.com/cockroachdb/cockroach/pkg/util/duration"
 	"github.com/cockroachdb/errors"
 )
 
@@ -98,191 +91,6 @@ func GetWindowFuncIdx(funcName string) (int32, error) {
 		return 0, errors.Errorf("unknown window function %s", funcName)
 	}
 	return funcIdx, nil
-}
-
-func (spec *WindowerSpec_Frame_Mode) initFromAST(w treewindow.WindowFrameMode) error {
-	switch w {
-	case treewindow.RANGE:
-		*spec = WindowerSpec_Frame_RANGE
-	case treewindow.ROWS:
-		*spec = WindowerSpec_Frame_ROWS
-	case treewindow.GROUPS:
-		*spec = WindowerSpec_Frame_GROUPS
-	default:
-		return errors.AssertionFailedf("unexpected WindowFrameMode")
-	}
-	return nil
-}
-
-func (spec *WindowerSpec_Frame_BoundType) initFromAST(bt treewindow.WindowFrameBoundType) error {
-	switch bt {
-	case treewindow.UnboundedPreceding:
-		*spec = WindowerSpec_Frame_UNBOUNDED_PRECEDING
-	case treewindow.OffsetPreceding:
-		*spec = WindowerSpec_Frame_OFFSET_PRECEDING
-	case treewindow.CurrentRow:
-		*spec = WindowerSpec_Frame_CURRENT_ROW
-	case treewindow.OffsetFollowing:
-		*spec = WindowerSpec_Frame_OFFSET_FOLLOWING
-	case treewindow.UnboundedFollowing:
-		*spec = WindowerSpec_Frame_UNBOUNDED_FOLLOWING
-	default:
-		return errors.AssertionFailedf("unexpected WindowFrameBoundType")
-	}
-	return nil
-}
-
-func (spec *WindowerSpec_Frame_Exclusion) initFromAST(e treewindow.WindowFrameExclusion) error {
-	switch e {
-	case treewindow.NoExclusion:
-		*spec = WindowerSpec_Frame_NO_EXCLUSION
-	case treewindow.ExcludeCurrentRow:
-		*spec = WindowerSpec_Frame_EXCLUDE_CURRENT_ROW
-	case treewindow.ExcludeGroup:
-		*spec = WindowerSpec_Frame_EXCLUDE_GROUP
-	case treewindow.ExcludeTies:
-		*spec = WindowerSpec_Frame_EXCLUDE_TIES
-	default:
-		return errors.AssertionFailedf("unexpected WindowerFrameExclusion")
-	}
-	return nil
-}
-
-// If offset exprs are present, we evaluate them and save the encoded results
-// in the spec.
-func (spec *WindowerSpec_Frame_Bounds) initFromAST(
-	ctx context.Context,
-	b tree.WindowFrameBounds,
-	m treewindow.WindowFrameMode,
-	evalCtx *eval.Context,
-) error {
-	if b.StartBound == nil {
-		return errors.Errorf("unexpected: Start Bound is nil")
-	}
-	spec.Start = WindowerSpec_Frame_Bound{}
-	if err := spec.Start.BoundType.initFromAST(b.StartBound.BoundType); err != nil {
-		return err
-	}
-	if b.StartBound.HasOffset() {
-		typedStartOffset := b.StartBound.OffsetExpr.(tree.TypedExpr)
-		dStartOffset, err := eval.Expr(ctx, evalCtx, typedStartOffset)
-		if err != nil {
-			return err
-		}
-		if dStartOffset == tree.DNull {
-			return pgerror.Newf(pgcode.NullValueNotAllowed, "frame starting offset must not be null")
-		}
-		switch m {
-		case treewindow.ROWS:
-			startOffset := int64(tree.MustBeDInt(dStartOffset))
-			if startOffset < 0 {
-				return pgerror.Newf(pgcode.InvalidWindowFrameOffset, "frame starting offset must not be negative")
-			}
-			spec.Start.IntOffset = uint64(startOffset)
-		case treewindow.RANGE:
-			if neg, err := isNegative(ctx, evalCtx, dStartOffset); err != nil {
-				return err
-			} else if neg {
-				return pgerror.Newf(pgcode.InvalidWindowFrameOffset, "invalid preceding or following size in window function")
-			}
-			typ := dStartOffset.ResolvedType()
-			spec.Start.OffsetType = DatumInfo{Encoding: catenumpb.DatumEncoding_VALUE, Type: typ}
-			var buf []byte
-			var a tree.DatumAlloc
-			datum := rowenc.DatumToEncDatum(typ, dStartOffset)
-			buf, err = datum.Encode(typ, &a, catenumpb.DatumEncoding_VALUE, buf)
-			if err != nil {
-				return err
-			}
-			spec.Start.TypedOffset = buf
-		case treewindow.GROUPS:
-			startOffset := int64(tree.MustBeDInt(dStartOffset))
-			if startOffset < 0 {
-				return pgerror.Newf(pgcode.InvalidWindowFrameOffset, "frame starting offset must not be negative")
-			}
-			spec.Start.IntOffset = uint64(startOffset)
-		}
-	}
-
-	if b.EndBound != nil {
-		spec.End = &WindowerSpec_Frame_Bound{}
-		if err := spec.End.BoundType.initFromAST(b.EndBound.BoundType); err != nil {
-			return err
-		}
-		if b.EndBound.HasOffset() {
-			typedEndOffset := b.EndBound.OffsetExpr.(tree.TypedExpr)
-			dEndOffset, err := eval.Expr(ctx, evalCtx, typedEndOffset)
-			if err != nil {
-				return err
-			}
-			if dEndOffset == tree.DNull {
-				return pgerror.Newf(pgcode.NullValueNotAllowed, "frame ending offset must not be null")
-			}
-			switch m {
-			case treewindow.ROWS:
-				endOffset := int64(tree.MustBeDInt(dEndOffset))
-				if endOffset < 0 {
-					return pgerror.Newf(pgcode.InvalidWindowFrameOffset, "frame ending offset must not be negative")
-				}
-				spec.End.IntOffset = uint64(endOffset)
-			case treewindow.RANGE:
-				if neg, err := isNegative(ctx, evalCtx, dEndOffset); err != nil {
-					return err
-				} else if neg {
-					return pgerror.Newf(pgcode.InvalidWindowFrameOffset, "invalid preceding or following size in window function")
-				}
-				typ := dEndOffset.ResolvedType()
-				spec.End.OffsetType = DatumInfo{Encoding: catenumpb.DatumEncoding_VALUE, Type: typ}
-				var buf []byte
-				var a tree.DatumAlloc
-				datum := rowenc.DatumToEncDatum(typ, dEndOffset)
-				buf, err = datum.Encode(typ, &a, catenumpb.DatumEncoding_VALUE, buf)
-				if err != nil {
-					return err
-				}
-				spec.End.TypedOffset = buf
-			case treewindow.GROUPS:
-				endOffset := int64(tree.MustBeDInt(dEndOffset))
-				if endOffset < 0 {
-					return pgerror.Newf(pgcode.InvalidWindowFrameOffset, "frame ending offset must not be negative")
-				}
-				spec.End.IntOffset = uint64(endOffset)
-			}
-		}
-	}
-
-	return nil
-}
-
-// isNegative returns whether offset is negative.
-func isNegative(ctx context.Context, evalCtx *eval.Context, offset tree.Datum) (bool, error) {
-	switch o := offset.(type) {
-	case *tree.DInt:
-		return *o < 0, nil
-	case *tree.DDecimal:
-		return o.Negative, nil
-	case *tree.DFloat:
-		return *o < 0, nil
-	case *tree.DInterval:
-		cmp, err := o.Compare(ctx, evalCtx, &tree.DInterval{Duration: duration.Duration{}})
-		return cmp < 0, err
-	default:
-		panic("unexpected offset type")
-	}
-}
-
-// InitFromAST initializes the spec based on tree.WindowFrame. It will evaluate
-// offset expressions if present in the frame.
-func (spec *WindowerSpec_Frame) InitFromAST(
-	ctx context.Context, f *tree.WindowFrame, evalCtx *eval.Context,
-) error {
-	if err := spec.Mode.initFromAST(f.Mode); err != nil {
-		return err
-	}
-	if err := spec.Exclusion.initFromAST(f.Exclusion); err != nil {
-		return err
-	}
-	return spec.Bounds.initFromAST(ctx, f.Bounds, f.Mode, evalCtx)
 }
 
 func (spec WindowerSpec_Frame_Mode) convertToAST() (treewindow.WindowFrameMode, error) {
