+++ pkg/sql/rowexec/BUILD.bazel
@@ -62,6 +62,7 @@ go_library(
         "//pkg/sql/catalog/typedesc",
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfra/execagg",
+        "//pkg/sql/execinfra/execexpr",
         "//pkg/sql/execinfra/execopnode",
         "//pkg/sql/execinfra/execreleasable",
         "//pkg/sql/execinfrapb",
