+++ pkg/sql/rowexec/filterer.go
@@ -9,6 +9,7 @@ import (
 	"context"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
+	"github.com/cockroachdb/cockroach/pkg/sql/execinfra/execexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra/execopnode"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/execstats"
@@ -23,7 +24,7 @@ type filtererProcessor struct {
 	execinfra.ProcessorBase
 	evalCtx *eval.Context
 	input   execinfra.RowSource
-	filter  execinfrapb.ExprHelper
+	filter  execexpr.Helper
 }
 
 var _ execinfra.Processor = &filtererProcessor{}
@@ -42,7 +43,7 @@ func newFiltererProcessor(
 ) (*filtererProcessor, error) {
 	f := &filtererProcessor{
 		// Make a copy of the eval context since we're going to pass it to the
-		// ExprHelper later (which might modify it).
+		// execexpr.Helper later (which might modify it).
 		evalCtx: flowCtx.NewEvalCtx(),
 		input:   input,
 	}
